Revert "[BROKEN] start breaking utils into multiple files"
This reverts commit 67cda36591fb1bf68e063140bcb43d28d57f99d2.
